requirements.txt
Authlib==1.5.0
certifi==2024.8.30
cloudinary==1.41.0
cryptography==44.0.1
fastapi==0.115.8
pydantic==2.10.6
pydantic_settings==2.8.0
PyJWT==2.10.1
pymongo==4.11.1
python-dotenv==1.0.1
pydantic_settings